Add setup image function to UIImageView extension
